Simplify worker / parser with async and modules
With the update to browserslist such that async / await is natively
supported, we can leverage async to simplify the try catch logic for
fetching and compiling the wasm module.

The parser is also simplified by importing the functions directly from
the wasm package instead of an intermediate object. This had confused me
initially, but the functions contain a reference to the global instance
local to the wasm package.