Overhaul wasm and js bridge
Have the network wasm endpoints directly return a `Vec<u8>`, which turns
into a Uint8Array instead of a String. The string for network data is a
wasteful middleman as wasm-bindgen converts the string to a byte array,
and since the blob which downloads the file doesn't care if it is a
string or an array buffer -- skip the intermediate string step.

Have the web worker use the byte array returned by the network wasm
endpoints and transfer it directly to the main UI thread instead of
needing a copy. This takes advantage of the transferable nature of an
`ArrayBuffer`: https://developer.mozilla.org/en-US/docs/Web/API/Transferable

Rewrite all wasm endpoints to return a Result so that we can remove the
heuristic on the JS side to try and detect if an error was embedded in
the returned string by detecting it's length.

Finally add tests to help against possible future regressions.

Overall, performance didn't seem to drastically change, but at least the
code is cleaner and tested.